requirements.txt
numpy-1.16.3
matplotlib
networkx-2.3
scipy-1.2.1